pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.1.5",
    "django-bootstrap5>=24.3",
    "django-filter>=24.3",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "gunicorn>=23.0.0",
    "ruff>=0.9.1",
]
